fix(kubernetes): restrict watch to server-side auth providers only
Client-side auth providers (google, oidc, aks) pass credentials from
the user's browser session, making long-lived watch connections
unreliable as tokens expire. Reject these providers with a BAD_REQUEST
error and log a warning.

Server-side providers (serviceAccount, googleServiceAccount, aws,
azure, localKubectlProxy) are unaffected.

refactor(kubernetes): extract KubernetesWatcher into separate class

Separate watch functionality from KubernetesFetcher into a dedicated
KubernetesWatcher interface and KubernetesClientBasedWatcher class.
Watching is a long-lived streaming connection that only works with
server-side auth providers, making it a distinct concern from the
request-response fetch pattern.

Extract shared connection logic (agent caching, in-cluster detection,
auth header building) into KubernetesConnection, shared between the
fetcher and watcher for connection reuse.
