+++ .changeset/watch-functionality.md
@@ -4,4 +4,4 @@
 '@backstage/plugin-kubernetes-backend': patch
 ---
 
-Add watch functionality to Kubernetes plugin REST client. The `watchResource()` method provides an async iterator interface for streaming resource changes from the Kubernetes API, supporting all event types (ADDED, MODIFIED, DELETED, BOOKMARK, ERROR) with the same error handling patterns as existing get/list operations.
+Add `KubernetesWatcher` interface for streaming Kubernetes resource changes via an async iterator. The watcher is separated from `KubernetesFetcher` because watching is a long-lived streaming connection that only works with server-side auth providers. Watch supports all event types (ADDED, MODIFIED, DELETED, BOOKMARK, ERROR) with errors yielded as data rather than thrown.
